project cleanup

+++ src/motifs_frequents_et_regles/Regle.java
@@ -10,16 +10,16 @@ public class Regle {
 	Double confidence;
 	Double lift;
 	public static int POSITIVE = 1, NEGATIVE = -1;
-	
-	
-	
+
+
+
 	public Regle(Ensemble<String> gauche, Ensemble<String> droite, Double confidence, double lift) {
 		this.gauche = gauche;
 		this.droite = droite;
 		this.confidence = confidence;
 		this.lift = lift;
 	}
-	
+
 
 	public static Regles regles_correlation(Regles regles, int signe) {
 		assert signe == POSITIVE || signe == NEGATIVE : "le signe doit être soit 1 pour les règles positives ou -1 pour les règles négatives";
@@ -31,7 +31,7 @@ public class Regle {
 		}
 		return regles_corr;
 	}
-	
+
 	public static Regles regles_correlation(Regles regles) {
 		Regles regles_corr = new Regles(regles.nbr_totale_instances);
 		for(Regle regle : regles) {
@@ -51,7 +51,7 @@ public class Regle {
 	public static Regles regles_association(Itemsets L, double min_conf) {
 		Regles regles = new Regles(L.nbr_totale_instances);
 		Supports supports = new Supports();
-		Itemsets L_multiples = new Itemsets(); 
+		Itemsets L_multiples = new Itemsets();
 		// prendre les itemsets multiples (taille > 1)
 		for(Ensemble<String> itemset : L) {
 			supports.ajouter(itemset);
@@ -59,7 +59,7 @@ public class Regle {
 				L_multiples.add(itemset);
 			}
 		}
-		
+
 		// former les regles associations depuis L_mult, en ignorant celles avec conf < min_conf
 		for(Ensemble<String> itemset : L_multiples) {
 			ArrayList<Ensemble<String>> sousEnsembles = itemset.sousEnsembles();
@@ -71,23 +71,23 @@ public class Regle {
 				double conf = sup_A_B / sup_A;
 				if(conf >= min_conf) {
 					A.support = (int) sup_A; // enregistrer les supports des ensembles, utile pour calculer les règles de corrélation
-					B.support = (int) supports.get(B);
+					B.support = supports.get(B);
 					double lift = conf / B.support * regles.nbr_totale_instances;
 					regles.add(new Regle(A, B, conf, lift));
 				}
 			}
 		}
-		
+
 		return regles;
 	}
-	
-	
-	
+
+
+
 	@Override
 	public String toString() {
 		return gauche.toString_simple() + "\t==>\t" + droite.toString_simple() + "\t(conf="+Dataset.arrondi(confidence)+"\tlift="+Dataset.arrondi(lift)+"\tsupp="+(int)(confidence * gauche.support)+")";
 	}
-	
+
 	public static class Regles extends ArrayList<Regle>{
 		private static final long serialVersionUID = 1L;
 		public int nbr_totale_instances;
@@ -95,7 +95,7 @@ public class Regle {
 		public Regles(int nbr_totale_instances) {
 			this.nbr_totale_instances = nbr_totale_instances;
 		}
-		
+
 		@Override
 		public String toString() {
 			String text = "";
